lambdas/notify/requirements.txt
requests===2.23.0
snowflake-connector-python
pandas